markit-backend/requirements.txt
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-braces==1.13.0
django-cors-headers==3.1.1
django-email-utils==1.0.0
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-email-auth==2.0.2
django-stdimage==5.0.1
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.0
progressbar2==3.47.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
PySocks==1.7.1
python-utils==2.3.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rest-social-auth==2.2.0
rope==0.14.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tweepy==3.8.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
wrapt==1.11.2
